# Alternative method to avoid rand contention in highly parallel usage