@mroth 2023-08-13
For people coming across this PR in 2023 and beyond: usage of `PickSource` is no longer required for this performance benefit when using go1.21 or greater, you can just use `Pick` as long as you aren't manually seeding global rand (which is no longer recommended). Changes documented in #28.